Add dockerfile env for the project
I'm still not quite sure this will be of some help, since, thinking of
a multi OS environment, I'm not sure of what tool I can rely on that
would launch the docker environment. Not even Make is available for
windows (and Windows is one of my base targets, along with linux).

Perhaps I could supply both a makefile and a powershell that will call
upon docker to do it's work.