Modify console pom to allow for testing with standard plugins in classpath
This one was tricky. To manually test the console client with the standard
plugins loaded, since one of them has a dependency on the console client
itself,it cannot be loaded by maven at build time (because of the cyclic
dependency between the console client and the plugin), and thus to make
its resources available (which means loading it in the classpath when
the application launches) we have to load the dependent plugins as extra
classpath elements with help from the exec:java plugin.