Remove all JMPS module code from the project
The ecosystem is not yet ready to support the new module system, and I've
lost quite a few days reading, trying and even encouraging dependencies
maintainers to migrate to a module system. But it is not an easy task
that will be done quickly.
We can revisit this in a few years and see what the situation is.

The code itself does not need mucho to be modularized. Only the deps are
a problem.