4XF-5 Add top menu bar fed from game commands
Adds a top bar menu to the top of the game window that is fed from the
list of configured game commands loaded by both plugins and the terminal
client.

The menus are accessed by pressing Ctrl+[menu shortcut key] and they
open a action list dialog that allows the user to navigate the menu and
its submenus.