4XF-26 - Create civilization screen
Add a create civilization screen opened from the select civilization
panel of the start game screen.

Also enables the capacity of the framework to save a civilization and
load it when loading plugins, saving the user generated civilizations
into the user's home 4xframework folder.

Consolidate xml as the configuration language and expanded the parsing
capabilities of the framework.