Creation of a new plugin to create cell type mappings in the correct place
The concept of cell type mapping goes to the console-client-api plugin, because it is a
concern only of the terminal and it is used by plugins depending on the console client.

The first definition of cell type mappings can be found in the civilization terminal client
adapter, which defines the mapping for the cell types defined in the civilization plugin.